Dead code elimination and fix edit bug
Attempting to edit the same event twice would not fill in the fields
properly. I am still not completely sure of the cause but I strongly
suspect it has to do with the fact that SvelteKit resets the form on
submission, but the variables the inputs were binding to remained the
same. So when the user tried to edit the same event, SvelteKit would not
detect a changed variable and update that input. This was fixed by
nulling out the variable(s) on edit submission.