Make schedule editing work without JS and fix bugs
In rare instances, an event would appear to be duplicated after saving
an edit. This was caused by the form action saving the new event before
the old one was deleted. This commit fixes this issue by introducing an
endpoint to edit an event by ID.

This commit also fixes the issue where if you clicked edit on an event,
then deleted it, then saved the edit, then clicked edit again, the form
wouldn't fill out, which was caused by not resetting the form's backing
data on save.

This commit also fixes some toasts displaying [object Object].