Password reset prompt now works between refreshes
If a user opened a password reset link and accidentally refreshed it in
the old design, that link would no longer work. That is because Lucia
Auth does not provide a way to check a token's validity without
consuming it. This new design optimistically assumes a link with a token
is valid. However, this means that a user trying to use an expired link
will only find out after submitting their new password. I believe this
is an acceptable tradeoff.