Lazy load statistics on users page
I believe the reason why the user page feels slower (on my old laptop at
least) is because it is loading every single word that occurs in a
sentence or paragraph question. We need to limit it to something like
the top 100, but for now this also mitigates the problem.