Fix preselection of current day on schedule
Previously, the schedule page would preselect tomorrow if it was past
noon today since it would choose the closest day to the current time and
tomorrow midnight is closer than midnight today after noon. To work
around this, set the current time to always be midnight for the purpose
of the preselection logic.